requirements.txt
Django>=5.2
ebooklib>=0.18
beautifulsoup4>=4.12.0
lxml>=4.9.0
Pillow>=10.0.0
markdown>=3.4.0
html2text>=2020.1.16
python-magic>=0.4.27
openai>=1.0.0
tiktoken>=0.5.0
requests>=2.31.0
channels>=4.0.0
daphne>=4.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
django-q2>=1.5.0
PyMuPDF>=1.23.0
